+++ srcjs/component/utils.js
@@ -13,8 +13,13 @@ import { SliderInput } from './SliderInput';
 import { Button } from './Button';
 import { Output } from './Output';
 import { Tab } from './Tab';
+import { Tabset } from './Tabset';
 import { Box } from './Box';
+import { UserBox } from './UserBox';
 import { ValueBox } from './ValueBox';
+import { InfoBox } from './InfoBox';
+import { BlockQuote } from './BlockQuote';
+import { Callout } from './Callout';
 
 export function getComponent(name) {
     if (name === "header") {
@@ -47,12 +52,21 @@ export function getComponent(name) {
         return new Output();
     } else if (name === "tab_panel") {
         return new Tab();
+    } else if (name === "tabset") {
+        return new Tabset();
     } else if (name === "box") {
         return new Box();
+    } else if (name === "user_box") {
+        return new UserBox();
     } else if (name === "value_box") {
         return new ValueBox();
+    } else if (name === "info_box") {
+        return new InfoBox();
+    } else if (name === "quote") {
+        return new BlockQuote();
+    } else if (name === "callout") {
+        return new Callout();
     }
-    
 
     return new Header();
 }