+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * {bs4Dash} page and components added to the available options (#31)
 
+* Elements can be deleted by right clicking an element and deleting (#35)
+
 * JavaScript code has moved to separate folder to work on as a module rather than 2/3 large files
 
 ### Bug Fixes